Rework ServletFilter to avoid duplicate providers
